requirements.txt
streamlit
torch
opencv-python-headless
pillow
matplotlib
numpy
scikit-learn
joblib